requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
mssql-django==1.3
pycparser==2.21
PyJWT==2.8.0
pyodbc==4.0.39
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
djangorestframework-simplejwt